chore: 파일 구조 변경 및 등록 관련 타입 지정
