+++ src/components/account/index.ts
@@ -3,6 +3,8 @@ import TabsNavigation from './TabsNavigation';
 import AccountBookDailySum from './AccountBookDailySum';
 import AccountBookDailyItem from './AccountBookDailyItem';
 import AccountBookDailyCard from './AccountBookDailyCard';
+import AccountForm from './AccountForm';
+import CategoryBox from './CategoryBox';
 
 export {
   TabsDisplayAccountSum,
@@ -10,4 +12,6 @@ export {
   AccountBookDailySum,
   AccountBookDailyItem,
   AccountBookDailyCard,
+  AccountForm,
+  CategoryBox,
 };
