+++ src/models/Account.ts
@@ -1,0 +1,6 @@
+export type CreateAccountForm = {
+  registerDate: string;
+  amount: string;
+  userCategoryId: string;
+  content?: string;
+};
