+++ src/models/Category.ts
@@ -1,0 +1,7 @@
+export type CategoryType = 'income' | 'expenditure';
+
+export type Category = {
+  id: string;
+  name: string;
+  categoryType: CategoryType;
+};
