+++ src/models/index.ts
@@ -1,3 +1,6 @@
 import User from './User';
+import { Category } from './Category';
+import { CreateAccountForm } from './Account';
 
 export { User };
+export type { Category, CreateAccountForm };
